move mobile nav logic to global hydration module
